+++ src/mympd_autoconf.c
@@ -76,7 +76,7 @@ sds get_mpd_conf(const char *key, const char *default_value) {
                 else if (strcasecmp(name, key) == 0 && strcasecmp(name, "password") == 0) {
                     sds *pwtokens;
                     int count;
-                    pwtokens = sdssplitlen(value, strlen(value), "@", 1, &count);
+                    pwtokens = sdssplitlen(value, (int)strlen(value), "@", 1, &count);
                     if (count == 2) {
                         if (sdslen(last_value) == 0 || strstr(pwtokens[1], "admin") != NULL) {
                             //use prefered the entry with admin privileges or as fallback the first entry
